Add notifications and dividend sync updates
Introduce a user notifications feature and integrate it into the dividend sync flow. Adds Notification entity, DTO, repository, service and controller with endpoints (/api/notifications/unread, /{id}/read, /read-all), plus a DB migration to create the notifications table. Update MasDividendosService to inject PortfolioRepository and NotificationService, validate/clean amounts, notify relevant users for announced dividends (approximate calculations per dividend type), and rethrow exceptions after logging. Change DividendAnnouncement fields to have defaults and switch id type to Int (update repository generic accordingly). Add default values to SyncState, add PortfolioRepository.findByStockSymbol, and tweak Quartz trigger startNow usage and scheduling comments.

+++ finanzas-back/src/main/kotlin/com/finanzas/dash/finanzas/repository/DividendAnnouncementRepository.kt
@@ -3,4 +3,4 @@ package com.finanzas.dash.finanzas.repository
 import com.finanzas.dash.finanzas.entity.DividendAnnouncement
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface DividendAnnouncementRepository : JpaRepository<DividendAnnouncement, Long>
+interface DividendAnnouncementRepository : JpaRepository<DividendAnnouncement, Int>
